changed the datatype for model and entity

+++ Excel/src/main/java/com/suraj/Excel/entity/ExcelEntity.java
@@ -10,17 +10,14 @@ public class ExcelEntity {
 	@Id
 	private Integer srl;
 	private String Datee;
-	private Long v_count;
-	private Long v_sum;
-	private Long db_cnt;
-	private Long db_sum;
-	private Long dp_count;
-	private Long dp_sum;
-	private Long load_trk;
-	private Long load_trk_off;
-	
-	
-	
+	private Integer v_count;
+	private Integer v_sum;
+	private Integer db_cnt;
+	private Integer db_sum;
+	private Integer dp_count;
+	private Integer dp_sum;
+	private Integer load_trk;
+	private Integer load_trk_off;
 	public Integer getSrl() {
 		return srl;
 	}
@@ -33,54 +30,56 @@ public class ExcelEntity {
 	public void setDatee(String datee) {
 		Datee = datee;
 	}
-	public Long getV_count() {
+	public Integer getV_count() {
 		return v_count;
 	}
-	public void setV_count(Long v_count) {
+	public void setV_count(Integer v_count) {
 		this.v_count = v_count;
 	}
-	public Long getV_sum() {
+	public Integer getV_sum() {
 		return v_sum;
 	}
-	public void setV_sum(Long v_sum) {
+	public void setV_sum(Integer v_sum) {
 		this.v_sum = v_sum;
 	}
-	public Long getDb_cnt() {
+	public Integer getDb_cnt() {
 		return db_cnt;
 	}
-	public void setDb_cnt(Long db_cnt) {
+	public void setDb_cnt(Integer db_cnt) {
 		this.db_cnt = db_cnt;
 	}
-	public Long getDb_sum() {
+	public Integer getDb_sum() {
 		return db_sum;
 	}
-	public void setDb_sum(Long db_sum) {
+	public void setDb_sum(Integer db_sum) {
 		this.db_sum = db_sum;
 	}
-	public Long getDp_count() {
+	public Integer getDp_count() {
 		return dp_count;
 	}
-	public void setDp_count(Long dp_count) {
+	public void setDp_count(Integer dp_count) {
 		this.dp_count = dp_count;
 	}
-	public Long getDp_sum() {
+	public Integer getDp_sum() {
 		return dp_sum;
 	}
-	public void setDp_sum(Long dp_sum) {
+	public void setDp_sum(Integer dp_sum) {
 		this.dp_sum = dp_sum;
 	}
-	public Long getLoad_trk() {
+	public Integer getLoad_trk() {
 		return load_trk;
 	}
-	public void setLoad_trk(Long load_trk) {
+	public void setLoad_trk(Integer load_trk) {
 		this.load_trk = load_trk;
 	}
-	public Long getLoad_trk_off() {
+	public Integer getLoad_trk_off() {
 		return load_trk_off;
 	}
-	public void setLoad_trk_off(Long load_trk_off) {
+	public void setLoad_trk_off(Integer load_trk_off) {
 		this.load_trk_off = load_trk_off;
 	}
 	
+	
+		
 		 
 }
